Database model initialisation and basic sync trial
